pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sbb"
version = "0.1.0"
description = "Self Building Box"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gymnasium[box2d,swig]>=1.2.0",
    "hyperopt>=0.2.7",
    "kagglehub>=0.3.13",
    "matplotlib>=3.10.5",
    "numpy>=2.3.1",
    "pandas>=2.3.2",
    "pytest>=8.4.1",
    "pyyaml>=6.0.2",
    "rustworkx>=0.16.0",
    "scipy>=1.16.1",
    "statsmodels>=0.14.5",
    "tabulate>=0.9.0",
    "tensorboard>=2.20.0",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
    "triton>=3.3.1; sys_platform == 'linux'",
    "yfinance>=0.2.66",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.mypy]
explicit_package_bases = true
namespace_packages = true
ignore_missing_imports = true
exclude = [
    ".mypy_cache/",
]
